[Dheer] Add: Implemented the funtionality find the best rated hotel for given date range for reward customer
